+++ Source/Coinbase.Pro/CoinbaseProClient.Deposits.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Coinbase.Pro.Models;
@@ -32,6 +33,24 @@ namespace Coinbase.Pro
       Task<CoinbaseDeposit> DepositFundsFromCoinbaseAccountAsync(
          string coinbaseAccountId, decimal amount, string currency,
          CancellationToken cancellationToken = default);
+
+      /// <summary>
+      /// Get information on a single deposit.
+      /// </summary>
+      Task<Deposit> GetDeposit(string transferId = null,
+         CancellationToken cancellationToken = default);
+
+      /// <summary>
+      /// Get a list of deposits from the profile of the API key, in descending order by created time.
+      /// </summary>
+      /// <param name="profileId">Limit list of deposits to this profile_id. By default, it retrieves deposits across all of the user's profiles</param>
+      /// <param name="limit">Truncate list to this many deposits, capped at 100. Default is 100.</param>
+      /// <param name="before">If before is set, then it returns deposits created after the before timestamp, sorted by oldest creation date</param>
+      /// <param name="after">If after is set, then it returns deposits created before the after timestamp, sorted by newest</param>
+      Task<List<Deposit>> ListDeposits(string profileId = null,
+         int? limit = null, DateTimeOffset? before = null, DateTimeOffset? after = null,
+         CancellationToken cancellationToken = default);
+
    }
 
    public partial class CoinbaseProClient : IDepositsEndpoint
@@ -75,6 +94,29 @@ namespace Coinbase.Pro
             .AppendPathSegment("coinbase-account")
             .PostJsonAsync(d, cancellationToken)
             .ReceiveJson<CoinbaseDeposit>();
+      }
+
+      Task<Deposit> IDepositsEndpoint.GetDeposit(string transferId,
+         CancellationToken cancellationToken)
+      {
+         return this.TransfersEndpoint
+            .WithClient(this)
+            .SetQueryParam("transfer_id", transferId)
+            .GetJsonAsync<Deposit>(cancellationToken);
+      }
+
+      Task<List<Deposit>> IDepositsEndpoint.ListDeposits(string profileId,
+         int? limit, DateTimeOffset? before, DateTimeOffset? after,
+         CancellationToken cancellationToken)
+      {
+         return this.TransfersEndpoint
+            .WithClient(this)
+            .SetQueryParam("type", "deposit")
+            .SetQueryParam("profile_id", profileId)
+            .SetQueryParam("before", before)
+            .SetQueryParam("after", after)
+            .SetQueryParam("limit", limit)
+            .GetJsonAsync<List<Deposit>>(cancellationToken);
       }
    }
 }