+++ HISTORY.md
@@ -1,3 +1,7 @@
+## v3.0.4
+* Added `Deposits.GetDeposit`
+* Added `DateTimeOffset` parameters to `Withdrawals.ListWithdrawals()` and `Deposits.ListDeposits()`.
+
 ## v3.0.3
 * Ensure `Withdrawals.ListWithdrawals()` returns `PagedResponse`.
 * Add `Withdrawals.GetWithdrawal()`.